Add WebXR controller support
