+++ package.json
@@ -48,7 +48,7 @@
     "present": "0.0.6",
     "promise-polyfill": "^3.1.0",
     "style-attr": "^1.0.2",
-    "three": "dmarcos/three.js#poseTarget",
+    "three": "dmarcos/three.js#webXRSupport",
     "three-bmfont-text": "^2.1.0",
     "webvr-polyfill": "^0.10.8"
   },
