+++ src/components/index.js
@@ -24,6 +24,8 @@ require('./shadow');
 require('./sound');
 require('./text');
 require('./tracked-controls');
+require('./tracked-controls-webvr');
+require('./tracked-controls-webxr');
 require('./visible');
 require('./vive-controls');
 require('./vive-focus-controls');
