+++ src/components/vive-controls.js
@@ -121,6 +121,7 @@ module.exports.Component = registerComponent('vive-controls', {
       idPrefix: GAMEPAD_ID_PREFIX,
       // Hand IDs: 0 = right, 1 = left, 2 = anything else.
       controller: data.hand === 'right' ? 0 : data.hand === 'left' ? 1 : 2,
+      hand: data.hand,
       orientationOffset: data.orientationOffset
     });
