+++ src/core/scene/a-scene.js
@@ -15,7 +15,7 @@ var initPostMessageAPI = require('./postMessage');
 var bind = utils.bind;
 var isIOS = utils.device.isIOS();
 var isMobile = utils.device.isMobile();
-var isWebXR = utils.device.isWebXR;
+var isWebXRAvailable = utils.device.isWebXRAvailable;
 var registerElement = re.registerElement;
 var warn = utils.debug('core:a-scene:warn');
 
@@ -40,6 +40,7 @@ module.exports.AScene = registerElement('a-scene', {
       value: function () {
         this.isIOS = isIOS;
         this.isMobile = isMobile;
+        this.hasWebXR = isWebXRAvailable;
         this.isScene = true;
         this.object3D = new THREE.Scene();
         var self = this;
@@ -262,14 +263,17 @@ module.exports.AScene = registerElement('a-scene', {
           vrDisplay = utils.device.getVRDisplay();
           vrManager.setDevice(vrDisplay);
           vrManager.enabled = true;
-          if (isWebXR()) {
+          if (this.hasWebXR) {
             if (this.xrSession) { this.xrSession.removeEventListener('end', this.exitVRBound); }
             vrDisplay.requestSession({immersive: true, exclusive: true}).then(enterXRSuccess);
           } else {
-            if (!vrDisplay.isPresenting) {
-              return vrDisplay.requestPresent([{source: this.canvas}]).then(enterVRSuccess, enterVRFailure);
+            if (vrDisplay.isPresenting) {
+              enterVRSuccess();
+              return Promise.resolve();
             }
+            return vrDisplay.requestPresent([{source: this.canvas}]).then(enterVRSuccess, enterVRFailure);
           }
+          return Promise.resolve();
         }
         enterVRSuccess();
         return Promise.resolve();
@@ -278,6 +282,9 @@ module.exports.AScene = registerElement('a-scene', {
           self.xrSession = xrSession;
           vrManager.setSession(xrSession);
           xrSession.addEventListener('end', self.exitVRBound);
+          xrSession.requestFrameOfReference('stage').then(function (frameOfReference) {
+            self.frameOfReference = frameOfReference;
+          });
           self.addState('vr-mode');
           self.emit('enter-vr', {target: self});
           // Lock to landscape orientation on mobile.
@@ -340,12 +347,11 @@ module.exports.AScene = registerElement('a-scene', {
         // Don't exit VR if not in VR.
         if (!this.is('vr-mode')) { return Promise.resolve('Not in VR.'); }
 
-        exitFullscreen();
         // Handle exiting VR if not yet already and in a headset or polyfill.
         if (this.checkHeadsetConnected() || this.isMobile) {
           vrManager.enabled = false;
           vrDisplay = utils.device.getVRDisplay();
-          if (isWebXR()) {
+          if (this.hasWebXR) {
             this.xrSession.removeEventListener('end', this.exitVRBound);
             this.xrSession.end();
             vrManager.setSession(null);
@@ -354,6 +360,8 @@ module.exports.AScene = registerElement('a-scene', {
               return vrDisplay.exitPresent().then(exitVRSuccess, exitVRFailure);
             }
           }
+        } else {
+          exitFullscreen();
         }
 
         // Handle exiting VR in all other cases (2D fullscreen, external exit VR event).
@@ -691,9 +699,10 @@ module.exports.AScene = registerElement('a-scene', {
      * Renders with request animation frame.
      */
     render: {
-      value: function () {
+      value: function (time, frame) {
         var renderer = this.renderer;
 
+        this.frame = frame;
         this.delta = this.clock.getDelta() * 1000;
         this.time = this.clock.elapsedTime * 1000;
