+++ src/systems/index.js
@@ -4,4 +4,5 @@ require('./gltf-model');
 require('./light');
 require('./material');
 require('./shadow');
-require('./tracked-controls');
+require('./tracked-controls-webvr');
+require('./tracked-controls-webxr');
