+++ tests/components/windows-motion-controls.test.js
@@ -71,7 +71,7 @@ suite('windows-motion-controls', function () {
       var removeEventListenersSpy = this.sinon.spy(component, 'removeEventListeners');
 
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -87,7 +87,7 @@ suite('windows-motion-controls', function () {
 
     test('does not detect presence of controller with missing id suffix', function () {
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = [
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = [
         {id: 'Spatial Controller (Spatial Interaction Source)', index: 0, hand: MOCKS.HAND_LEFT, pose: {}}
       ];
 
@@ -102,7 +102,7 @@ suite('windows-motion-controls', function () {
 
     test('does not detect presence of controller with unknown device ID', function () {
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = [
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = [
         {id: 'unknown', index: 0, hand: MOCKS.HAND_LEFT, pose: {}}
       ];
 
@@ -118,7 +118,7 @@ suite('windows-motion-controls', function () {
     test('does not detect presence of controller with wrong hand', function () {
       // Mock isControllerPresent to return false.
       component.data.hand = MOCKS.HAND_RIGHT;
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_LEFT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_LEFT);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -132,7 +132,7 @@ suite('windows-motion-controls', function () {
     test('does not detect presence of controller in second pair with not enough connected', function () {
       // Mock isControllerPresent to return false.
       component.data.pair = 1;
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -146,7 +146,7 @@ suite('windows-motion-controls', function () {
     test('detects presence of controller in third pair', function () {
       // Mock isControllerPresent to return true.
       component.data.pair = 2;
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT, MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT, MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT, MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT, MOCKS.HAND_LEFT, MOCKS.HAND_RIGHT);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -166,7 +166,7 @@ suite('windows-motion-controls', function () {
 
       function detect (hand) {
         component.data.hand = hand;
-        el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(hand, hand);
+        el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(hand, hand);
 
         // delete our previously created mock, so component behaves as if it's never
         // checked for controller presence previously.
@@ -187,7 +187,7 @@ suite('windows-motion-controls', function () {
 
       function detect (hand) {
         component.data.hand = hand;
-        el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList('', '', '', '');
+        el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList('', '', '', '');
 
         // delete our previously created mock, so component behaves as if it's never
         // checked for controller presence previously.
@@ -208,7 +208,7 @@ suite('windows-motion-controls', function () {
 
       function detect (hand) {
         component.data.hand = hand;
-        el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList('', '');
+        el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList('', '');
 
         // delete our previously created mock, so component behaves as if it's never
         // checked for controller presence previously.
@@ -224,7 +224,7 @@ suite('windows-motion-controls', function () {
       component.data.hand = MOCKS.HAND_RIGHT;
 
       // Mock isControllerPresent to return false.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -239,7 +239,7 @@ suite('windows-motion-controls', function () {
       component.data.hand = MOCKS.HAND_RIGHT;
 
       // Mock isControllerPresent to return false.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -254,7 +254,7 @@ suite('windows-motion-controls', function () {
       component.data.hand = MOCKS.HAND_LEFT;
 
       // Mock isControllerPresent to return false.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -269,7 +269,7 @@ suite('windows-motion-controls', function () {
       component.data.hand = MOCKS.HAND_LEFT;
 
       // Mock isControllerPresent to return false.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED, MOCKS.HAND_UNHANDED);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED, MOCKS.HAND_UNHANDED);
 
       // delete our previously created mock, so component behaves as if it's never
       // checked for controller presence previously.
@@ -286,7 +286,7 @@ suite('windows-motion-controls', function () {
       var removeEventListenersSpy = this.sinon.spy(component, 'removeEventListeners');
 
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
 
       // Mock to the state that a gamepad is present.
       component.controllerEventsActive = true;
@@ -318,7 +318,7 @@ suite('windows-motion-controls', function () {
 
   suite('axismove', function () {
     test('emits thumbstick moved on X', function (done) {
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
       // Do the check.
       component.checkIfControllerPresent();
       // Install event handler listening for thumbstickmoved.
@@ -333,7 +333,7 @@ suite('windows-motion-controls', function () {
     });
 
     test('emits thumbstick moved on Y', function (done) {
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
       // Do the check.
       component.checkIfControllerPresent();
       // Install event handler listening for thumbstickmoved.
@@ -348,7 +348,7 @@ suite('windows-motion-controls', function () {
     });
 
     test('emits trackpad moved on X', function (done) {
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
       // Do the check.
       component.checkIfControllerPresent();
       // Install event handler listening for trackpadmoved.
@@ -363,7 +363,7 @@ suite('windows-motion-controls', function () {
     });
 
     test('emits trackpad moved on Y', function (done) {
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
       // Do the check.
       component.checkIfControllerPresent();
       // Install event handler listening for trackpadmoved.
@@ -378,7 +378,7 @@ suite('windows-motion-controls', function () {
     });
 
     test('does not emit thumbstickmoved if axismove has no changes', function (done) {
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
       // Do the check.
       component.checkIfControllerPresent();
       // Fail purposely.
@@ -402,7 +402,7 @@ suite('windows-motion-controls', function () {
       component.createControllerModelUrl = function () { return TEST_URL_MODEL; };
 
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
 
       // Perform the test
       component.checkIfControllerPresent();
@@ -415,7 +415,7 @@ suite('windows-motion-controls', function () {
       var loadModelSpy = this.sinon.spy(component, 'loadModel');
 
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_DEFAULT);
 
       // Perform the test
       component.checkIfControllerPresent();
@@ -432,7 +432,7 @@ suite('windows-motion-controls', function () {
       component.data.hand = MOCKS.HAND_RIGHT;
 
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_RIGHT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_RIGHT);
 
       // Perform the test
       component.checkIfControllerPresent();
@@ -449,7 +449,7 @@ suite('windows-motion-controls', function () {
       component.data.hand = MOCKS.HAND_RIGHT;
 
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_UNHANDED);
 
       // Perform the test
       component.checkIfControllerPresent();
@@ -473,7 +473,7 @@ suite('windows-motion-controls', function () {
       };
 
       // Mock isControllerPresent to return true.
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_RIGHT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_RIGHT);
 
       // Perform the test
       component.checkIfControllerPresent();
@@ -505,7 +505,7 @@ suite('windows-motion-controls', function () {
 
     function buttonTestHelper (done, buttonIndex, buttonName) {
       var state = {value: 0.5, pressed: true, touched: true};
-      el.sceneEl.systems['tracked-controls'].controllers = createMotionControllersList(MOCKS.HAND_RIGHT);
+      el.sceneEl.systems['tracked-controls-webvr'].controllers = createMotionControllersList(MOCKS.HAND_RIGHT);
       // Do the check.
       component.checkIfControllerPresent();
       // Install event handler listening for changed event.
